Basic Skeleton of Monitoring logic written.
Have to verify the logic wrt flow and finetuning.

Have added Internet permission.

As the List of Channels to monitor can be independently updated by GUI,
independent of the monitoring thread, the monitoring thread now saves the
list of channels it started out with and uses that saved value thro its
logics.

NOTE: Currently the GUI can only add to the list and not modify or delete
so saving number of channels when the monitor logic started out is good
enough for now.

NOTE2: Maybe set a different color for items in the channel list which are
currently not being monitored (i.e wrt above scenario).