Exception cleanup
a) If taking the user entered data from the edittext widgets into the channel
monitor list raises a exception, now a toast message is shown to user.

This is in addition to the existing numOfItems toast, which also gives the
info indirectly by repeating the same value as last time.

b) Have put the onProgress logic inside a try-catch block, so that if there
is any failure wrt working with the gui elements (mainly), it is trapped
and a toast message shown to the user.

This would ensure that the race which we had previously wrt trying to update
the ui element corresponding to a channel in the monitor list, leading to a
exception, due to the channel not being part of the MCasts listview yet, will
now get handled as gracefully as possible at one level.