MCRedDelay - User decides when things get red hot
Previously if the delay count crosses 100, then the background of
MCastName of the channel was made red in the MCasts listview.

HOWEVER now, the user can specify the delay count for a given channel
being monitored, at which point it should be considered as sufficiently
delayed to make its line red.