SeqNumberOffset in MCastPacket
The user can now specify a offset within the channel packets, which
will be treated to contain a 32bit little-endian sequence number.

TODO:
THis will be used to get the seqNo from this offset for the packets
recieved from the given channel and then show the same as part of
the MCasts listview.