Show SeqNum of monitored channels in MCasts listview
The sequence number embedded in packets received on each monitored
channel, is extracted from the specified corresponding seq num offset
and inturn shown as part of the MCasts listview for the corresponding
channel.

The SeqNum is treated to be a 32bit value stored in little-endian
format, in the recieved packets.