Cleanup SeqNum logic
a) By default when the channel is added to monitored list, the SeqNum
is set to 0

b) If there is a exception when trying to get the seq number, it is
set to a predefined value 987654321

c) Logic used to get the SeqNum from the packet/buffer is configured
to treat the embedded data to be in LittleEndian byte order.

NOTE: If the data recieved is smaller than the amount required to
extract the SeqNum from the packet buffer, then any old data in the
packet buffer will also get mixed up. This is because

a) The Packet buffer is reused across channels as well as across
packets of the same channel.

b) The packet buffer is not cleared between these uses.