Updated tracked seqNum only if larger
Even thou in my previous commit I had mentioned that the seqNum is updated only
if the new seqNum got in the packet is larger than what was seen already, the
actual code had not been updated for the same, so now I have updated the logic
to match the same.

NOTE: On a related note, the reason one of the counters for seqNum tracking is
called OlderSeqs and not RepeatedSeqs, is because, we don't maintain a list of
all seen seqNums and then check each new seqNum got against this list.

With the current logic we can only know if the new seqNumber got is older(i.e
smaller or same, assuming a ascending seqNum) than what is expected