0th MCasts lv entry header, appropriate NumMCasts
* Now the 0th MCasts list view entry is treated has a header which tells the
meaning of each column/field.

To achieve this the 0th entry is skipped when adding channels to monitor to
Monitored MCasts list.

When creating the entries in the MCasts lv, the 0th entry is treated
specifically as a header entry. Also the default value of the fields is now set
to the field's name rather than a meaningful default value for that field; the
only exception being the Name and Info fields, which carry the field name for
the 0th entry and actual data for the subsequent entries.

The Async task skips the 0th entry of the monitored MCast channels list. As it
is not used, to keep in sync with the 0th entry in the MCasts lv being a
header.

Now when deciding whether to update progress each time thro the loop or once
every 10 iterations thro the packet monitoring loop, it is checked against 2
instead of 1, because we have to account for the dummy header entry also.

The above changes are done to handle the 0th entry being a header row in a
simple way.

* In the Monitoring async task, at one place, I had wrongly used the global
iNumOfMCasts directly instead of using the value saved at the begining of the
task. This has been fixed now.

This mistake had creeped in when I had added logic to check if only one channel
is being monitored, in which case assume the program is monitoring a high
throughput channel and update progress in gui only once every 10 iterations
thro the monitoring loop.