GUI Monitor Progress update once every 5 secs for high throughput mode
To minimise packet losses due to GUI updations, now the gui is updated wrt
progress updates only once every 5 seconds, instead of once every 0.5 seconds.
NOTE THAT this logic is only applicable when only a single mcast channel is
being monitored, in which case the logic currently automatically assumes that
it could be a high throughput channel.