NwTest: Use select to achieve specified throughput
For the internally specified throughput, if the actual time used for sending a
bunch of packets is lesser than the time alloted for the same, then select is
used to sleep off the time remaining, so that we maintain the specified
throughput.

NOTE:
Currently the previous brute force and stupid throttling logic of sleeping for
1 second, every N*N packets, leads to a small goof-up with the throughput
specified, once every N*N packets.