ProgressUpdate: Almost Exactly once every 5 seconds
Instead of using a indirect representation of passing of time, by using a count of number of iterations thro the monitor logic loop, now the time is explicitly read with almost milisecond precision (provided the system supports the same) and used to decide if 5 seconds have occured.

Thus irrespective of the packet reception timesout or packets are read, it will accurately track time and thus progress is updated on the GUI, once every 5 seconds.