Increase commonality btw JavaNet&JavaNIO NwInterfaces
Have added the logic which displayes available network interfaces in the
target, from the Java.NIO to Java.Net based logic.

THis is to ensure that the difference between both the branches are as minimal
as possible. Technically currently java.net based logic doesn't use network
interfaces related info, but at same time there is no harm in having the same.