Basic DataHandler skelton - Only Logging logic now
Have added a DataHandler class, which will be used for handling of packets
recieved over the network. Currently

a) It opens the data and the log files

b) Function to write lost packets data to the log file.

The MainActivity currently uses the Applications ExternalStorage Directory as
the base path into which the data and log files are stored.