Optimize Logging logic - Use Queues - P2b
As the Thread deprecates the use of stop/cancel call wrt the Thread, Instead a
special marker value is placed into the lost packets info queue, in turn the
run logic of the file based logging runnable will monitor for these markers and
quit when found.

Note that the special marker used for stopping the logging related thread logic
is 4321 for LostPacketStartSeq and 1234 for LostPacketEndSeq, and in-turn as
such a seq of lost packets can never occur in reality, this is a safe way of
passing the stop notification to the thread logic.

NOTE: With this queue based logic, with the Android api 22 Arm based virtual
machine, for a 3 MegaBytesPerSec channel, the number of lost packets compared
to recieved packets reduced from 33% to 20%. Have to check on a real target,
how things work out, hopefully it should be 0% i.e no packet loss because of
logging.