PacketData handling p1 - DH initial logic
Added a counting semaphore based initial Consumer skeleton to DataHandler
class. If it can release the semaphore, it will assume that data has been filed
in a related corresponding buffer. To avoid any race of the Producer writing to
the same buffer which the consumer is consuming the counting semaphore's count
is 1 or more less than the actual number of buffers that are available.
Currently the counting semaphore count is kept 3 less than actual number of
buffers.

Both the producer and consumer use their own indexes into the data buffer. This
data buffer is used as a circular buffer, which is used/interpreted as
consisting of predetermined fixed size blocks.

The DataHandler class contains the logic which is to be called by producer
thread, as well as the logic which is to be called by consumer thread.

The dataBuffer has space for 2048xMaxN blocks. However as the data sent over
the network is currently only 1024 in size, there is a dataSize field in DH,
which is used to set the actual used data block size i.e 1024 in this case.