PacketData saving p2 - BlockingQueue required beyond CountingSem
As I had doubted (bcas the sem.release wasn't throwing a Interrupted
exception), the Semaphores release logic doesn't block. Which means my old
logic wrt syncing producer and consumer using counting semaphore won't work.

So now I have added a blocking queue into the logic. With this now I can
reposition the release to be after the buffer is used, so that there can't be a
race condition like what I had feared previously (and in turn to take care of
avoiding that possible race, I had reduced the counting sem's count to be 1 or
more less than the number of buffers available - This delta is not required
now).