PacketData saving p3 - Thread->Runnable stopping
Added logic to control whether the loop wrt saving data bufs is active or not.
This is controlled through a boolean which is checked to verify as to whether
the SaveDataBufs logic will remain in loop or not.

When the thread is started, this control variable is set true, before calling
the SaveDataBufs function. ANd inturn when the Async task wrt channel
monitoring is stopped (be it for a proper reason or because of a error), this
variable is set to false. Inturn the SaveDataBufs loop will exit, helping the
thread to quit.