PacketData saving p4 - Simplified stopping & General
As the semaphore.release doesn't block, and in-turn BlockingQueue had to be
brought in, the old overall flow logic is no longer required. Also the flag
based stopping logic which I implemented in last commit wont work as queue.take
is a blocking logic and unless there is some data to read from the queue, the
logic will never get a chance to check if the loop should be continued or quit
thro the flag.

So

a) Have removed the semaphore usage fully. Because BlockingQueue itself allows
limiting/controlling access to the multiple logical blocks in the data buffer
in a controlled manner between the producer and consumer.

b) Also for stopping the save data buffers loop, instead of the flag in
DataHandler, now a special marker is passed from producer to the consumer over
the BlockingQueue.

If null is passed as the data buffer to Write2DataBuf function, it
automatically inserts the special marker into the Queue. Inturn the
SaveDataBufs logic on seeing the special marker will quit the loop.