DataHandler Open & Close Files
Now Opening of Log and Data files of the DH is moved into a seperate function
called OpenFiles, instead of being in the Constructor. Also a CloseFiles
function is implemented.

This ensures that the logic using DH can decide when to open and when to close
these files.

Inturn the MainActivity has been updated such that when StartMon is called,
first CloseFiles is called and inturn OpenFiles is called. So if a user stops
and starts monitor logic, the files will contain data and info wrt the new
StartMon activity.

Also onDestroy now calls CloseFiles just to be safe.