1st packet with SeqNum 0  handling, Msg cleanup
There was a corner case with the 1st packet with seqNum 0, being handled
wrongly as a OlderSeqs packet. This was because the LastSeenSeqNum related
array was being initialised to 0 rather than -1. This has been fixed now.

Also the LogTaskQ start and stop messages were wrongly logged as a warning
rather than info message, this has been fixed now.

When data_stop logic is called when monitoring async task finishes one way or
the other, the special marker message which is pushed to the blocking queue is
identified in log messages now.