publish progress at the end - partly
Tried to publish the progress in the gui by calling publishProgress when coming
out of the while loop in the doProcess loop. But that doesnt seem to be
helping.

So have now also updated the LogStrLn to log the messages it is given to logcat
also. So one can see the summary info (which is the only logic using LogStrLn
for now) not only in the log file but also in the logcat.