Update progress from Monitoring AsyncTask at end
Had forgotten the semantic/flow of AsyncTask. IE the reason why the logic was
not updating the monitoring tasks progress, in the GUI, at the end when the
Monitor task is stopped. Me calling publishProgress after exiting the
monitoring loop on cancel doesn't help as onProgressUpdate wont be called after
the task is cancelled.

So now I have removed this unwanted publishProgress call. As well as I have
moved the logic which updates the GUI wrt progress info into a seperate
function which is called from onProgressUpdate, onCancelled and onPostExecute.
So in all cases the GUI will be updated with almost the most recent progress
update.

The reason why I have called this almost and not necessarily the final progress
update, at one level is, because the storage thread will still be processing
messages in the blocking queue. But in this the almost and final status should
be same in almost all cases. The main exception being, if there is a transfer
active in the monitored channel, when stop->cancel was triggered, then the
actual nw data available on the target and the view of channel performance from
the monitoring logic perspective may not be fully up to date, as the monitoring
logic can get canceled even before all nw data is processed. ALso in this case,
saving and logging operations triggered from the monitoring task may have been
blocked waiting for space in the blocking queue, which will get interrupted
when the AsyncTask is cancelled.