DataRecovery p2 - PresenceInfo flushed out
**** Client side

* The initial address used to send presence info packets to help identify the
Server which can help with data recovery is no longer hardcoded. Instead by
default universal broadcast address of 255.255.255.255 is set, but it can be
changed to a new address if required.

The address 255.255.255.255 doesn't seem to work, even if broadcast is enabled
for the socket in question. Instead now the user is given the chance to
explicitly set the initial address to send the PI packets by entering the same
in the GUI.

For now temporarily the Sole Editbox in the Unicast section of GUI is used to
get the initial PI related address to query. TODO: Later I may ask user to
specify the interface to use for PI packets and inturn get the ip address
assigned to that interface and in turn the ip broadcast address to use.

* Replaced the Toasts in the PresenceInfo related runnable to logging.

* Few more log messages

**** Server side

Set a default timeout for the socket corresponding to 1 second

Maintain a clients list, which is updated when ever PI packets are recieved
during PI phase from new clients. Also for each recieved PI packet a PI
response is sent back to the client.

deltaTime calculated is in seconds, so updated the presence info related loops
exit condition appropriately to correspond to 10 minutes.

Exit at the end of PI phase.