PI use constants, Unicode-16 ID, list clients
On both Server and Client side now use predefined constants rather than hard
coded values directly in the code.

As the ID of the client is stored as a unicode-16 string, copying of the same
into the PIReqPacket as been updated to treat and retain it as unicode-16,
rather than as ascii. Previously wrongly interpreting the string as ascii value
had resulted in corrupted ID.

The PIReqPacket is now exactly 20 bytes in length, which consists of
* 4 bytes Special/Marker-SeqNum
* 8 unicode-16 chars i.e 16 bytes of ID

Server prints the recieved PI pkts from the client.

Server prints the list of clients at the end of PI phase.