Toggle whether MCast Data is saved or not by Monitor logic
A new MCastDataON/OFF button is added to gui. Before selecting StartMon, first
set MCastData button to MCastDataON or MCastDataOFF. Then when monitor logic is
started, it will either save the mcast data of the 1st channel being monitored
or ignore it.

NOTE that changing the MCastData button state in the middle of Monitoring
doesnt change anything. It is checked and internal state suitably setup only
when StartMon button is pressed.