Unicast Recovery Client - basic Skeleton
On the client side, in a loop, it keeps doing the following

a) If a URReqSeqNum is recieved, then FOR NOW a short list of current first 10
lost packet ranges is sent.

b) If some other Nw Command is recieved, it is ignored, along with a log entry.

c) If a data packet is recieved, it is ignored with a log entry

TO help with same,

Now the lostPackets RangeList in MainActivity is made into a static member.

getStart and getEnd functions added to RangeList, such that they give the range
start and range end values for the currently specified index into the RangeList