UnicastRecovery - Bit More 02
* Peer IP Address (InetAddress) compared using appropriate equals function.

* log_stop moved to Async task from UI thread

Saving of lostPackets RangeList takes non-deterministic time as it is a file
operation and the number of entries to write will depend on amount of
lostPackets to write about. So UI thread is wrong location to trigger it.

Also in doInBackground, the logic after while(!isCancelled) loop should be run
irrespective of whether the loop exits because of cancelling of the AsyncTask
or because of some predetermined condition, one breaks from the loop. So the
time consuming lostPackets saving to file moved here. NOTE: Care should be
taken while implementing logic such that one doesn't return from with-in the
loop.

Also on successful saving of lostPackets RangeList, a message is logged.

TODO: For now I have reduced PI phase time to 3 minutes from default 10
minutes. Have to revert back later.