UnicastRecovery - Quit thread on Program exit
Previously the thread which runs PI and UR, would have remained permanently in
UR phase, even if the application is exited by user.

TODO: Have to check if one explicitly clears the app/MainActivity from previous
activities list maintained by ANdroid, if the thread will be killed, or whether
it will persist.

Now when the application is exited and its onDestroy is called, the UR logic is
sent a message through a flag to indicate it to exit the loop.