UnicastRecovery - Endianess, unpacking
On Server side:

Fixed the logic to recover the response command and lost packets info block

On Client side:

I was wrongly interpreting the command in the nw packets as BigEndian, as
ByteBuffer by default is set to BigEndian. Have fixed it now by setting it
explicitly to LittleEndian.

TODO: Will have to change the other ByteBuffers used in recently added code on
client side to also be LittelEndian, later.