UnicastRecovery - Fix corner cases 01
ClientSide

* Fixed a oversight with copy paste from PI logic, where the response packet
size was set to the PI response buffer size. Now it is updated to correspond
to the length of UR response.

* RangeList was wrongly expanding its 1st entry to include the subsequent
entries, that were tried to be added to the list. As long as the new entries
were beyond the 1st entry's range. This creates a problem if the new entry is
fully beyond the 1st/current entry.

Now if the new entry is beyond the current entry being checked against, then
it will get added to immidiately follow the current entry.

TODO: Note that this logic is not proper, because, we should skip such a case
and inturn let the logic check if a subsequent entry/ies in the existing list
overlaps this new entry and or let all entries which are before this new
entry to be skipped before adding this new entry. This also means the same
logic can take care of the 1st entry added to the list. But currently the 1st
entry is handled as a special case, I require to merge 1st entry as well as
the case mentioned above.

For now for the normal packet loss case during multicast transfer phase,
except for the 1st entry all subsequent entries will get insert such that list
will contain them in descending order rather than ascending order.