Fix struct pack bytes interpretation goof up - p1
Have fixed the struct.pack used in the nw test scripts wrt the dummy test data
they generate for the data packets.

TODO: Other uses require to be fixed later.

NOTE: The help of struct, always was seeming to indicate this. But for some
reason when I initially implemented the logic, I vaguely remember to have read
the struct help to mean bytes will be included based on their actual size, if
explicit size is not prefixed in the format string of struct.pack for bytes.
So I ignored what help(struct) was seeming to tell differently. But after I
implemented the check-image script, I noticed that all locations in the saved
data file seemed to have only marker ids 0-255 repeating i.e 256th block in
data file had 0 instead of 256 and so on. That forced me to relook at this
issue and then inturn fix the bug (introduced by strange circumstance).