Added basic mcast skeleton for c based client
As the android java.net based gui logic doesn't seem to be recieving multicast
packets larger than 8 to 16 bytes, from my python nw-send-mcast script, which
is a high throughput mcast channel.

While at same time tcpdump seems to be showing the packets. I have held
Wifi.MulticastLock as specified by android documentation, but still it doesn't
seem to be helping. So am re-implimenting the client side logic in c here.

This contains the basic/initial skeleton for the mcast data transfer (capture
and save) logic. Here I have tried to create a mcast socket.

TODO: Previously the logic seemed to be picking slow throughput, but larger
mcast channel packets, have to recheck this once again. If they are still being
recieved, then have to check, if having low-latency+high-throughput QOS flag
while sending the packets, is helping in those cases, in which case, I will
have to add the same in my nw-send-mcast script.

TODO: Have noticied, that this code is failing at run time with no such device,
for the setsockopt call, when trying to join the mcast group. So have to
explicitly specify a ifindex and then see if it helps.