filedata stdout, filedata_save check return values
Now file to use for saving recieved data from the network, can be specified as
a file number in the commandline. This allows one to specify 1 as the filedata
file number in the commandline, which will result in the recieved data being
written into stdout. Inturn one is required to redirect stdout to a file in the
commandline.

TODO: Later will allow the user to specify a file name in the commandline and
inturn open a file handle into it.

Now the seek and write operations wrt the filedata are checked for errors and
suitable warning messages written to stderr.