Display Stats wrt pkt reception, usleep for 1sec
Added logic to track time wrt statistics display as well as the last time when
data was recieved.

Inturn once every predefined seconds has elapsed, mcast channel performance
related statistics is printed.

TODO: Have to check, if bind is necessarily required to recieve mcast packets
or not. Once I verify that, I have to add a bind to the mcast socket to bind it
to a particular port, so that I don't recieve packets related to non mcast file
transfer logic in this socket.

TODO: Have to check, if I can seek on stdout, if not, I have to add that
passing of file name to program wrt file to use for data saving as to be
implemented at a higher priority.