data file - use specified path+name to create/open
Previously I had taken file index of the process as the reference for where to
save the data. However with this I can only write to stderr or stdout (or its
redirections in the commandline). But with this as I had suspected one can't
seek, so doesn't help our case where there could be lost data inbetween and we
require to seek to a new location, by skipping inbetween space related to lost
packets.

So now I expect the user to specify a file name with path. In turn the file is
openned with create flag, so that it will get created if required. The file
will be openned in read+write mode. Also if the file is being created then
permissions are setup such that owner of the file as read and write access.

NOTE that currently I don't truncate the file, if it already exists. This
should also help the following case i.e if the user passes a block device file
as the data file to save to.

Fixed a mistake where I was checking the full expected packet size against the
size of data to save for a give packet, and inturn emitting a warning, if they
mismatch, which would have been the case always.

If one is not able to open the data file to save to, then the logic will emit a
warning and then continue to monitor the channel and print related statistics,
without trying to save the data if any recieved on this monitored channel.