MCastRecv: Update iPrevSeq
Had forgotten to update iPrevSeq and that was leading to strange results wrt
the periodically printed statistics, where both PktCnt and DisjointPktCnt was
increasing such that iPktCnt was in sync with SeqNum+1 (as expected if all
packets are recieved) while iDisjointPktCnt was growing faster each time.

Put a debug print to verify the same, and then fixed the oversight.