MCastRecv: mcastSocket init cleanup, exit on error
Now the port to bind to for the mcast socket is passed as a argument to the
mcast init function, instead of it directly taking it from the global variable.

On success wrt mcast socket creation, the init function itself prints a success
message with details, instead of the main program logic doing it.

Now even for bind failure wrt mcast socket, the program will exit. Keeping it
consistent wrt behaviour for other mcast socket initialisation failures.