C MCastLogic: cleanup, MULTICAST_ALL, Arg indexes
Command line arguments are now accessed by defining constants corresponding to
their positions in the commandline. This ensures that if new arguments are
added or some arguments repositioned or arguments removed, the logic doesnt
require to change much, instead just by redefining these constants, most of the
logic can remain unchanged.

Now the local address used as part of the mcast join setsockoption is also
taken from the commandline. With this to specify the local interface which
should be used to join a mcast group, remember to either

a) specify the proper interface index & inturn specify local address as 0.0.0.0

b) specify 0 for interface index and specify local address of the interface to
use for mcast joining.

Now the Multicast related socket is optionally configured wrt the socket option
of IP_MULTICAST_ALL. WHich seems to indicate that send mcast packets which
belong to the port which this socket has bound to, even if this socket didn't
explicitly join that particular mcast group specificially. Have to experiment
this more later. Have added this to test the strange behaviour found on some
android targets. TODO: May disable this by undef'ing the enabling define.

Have cleaned up the messages emitted by the program in general.

NOTE: For some strange reason, all the updates and commits I did from yesterday
afternoon 2:30 pm or so till late in the night, have gone missing in the
filesystem and inturn the git. So this is a recovered re-commit after doing a
raw search for strings in the disk in a raw way.