LinkedListRange - Add a context structure, print
The functions instead of directly working on global variables (rather mainly
the start node of the linked list), now get the context/start_node wrt the
linked list being operated on, from the context structure LLR passed to it.

This allows this library to support multiple independent linked lists to be
managed by it, parallely at the same time.