LinkedListRange - Delete, _alloc
Moved the logic which allocates a node and initialises it's members with
suitable values from a standalone perspective, into a new _ll_alloc function.

Inturn this is used by the existing ll_add_sorted and the new ll_delete logics.

Have added a delete logic, which takes a single value and then updates the list
of ranges such that this value is excluded from that list. This could involve
removing node if it represented this single value or else adjusting existing
ranges as required. If the val is the begin or end of a existing range, just
that begin/start or end needs to be suitably updated. However the given value
falls within a given range, then that range requires to be split into two new
ranges, such that this particular value is excluded.