LinkedListRange - _ll_add, add_sorted uses it
Added a generic _ll_add logic, which takes the current position/node after
which a given new node should be added.

It takes care of the following cases:

a) As the first node in a new ll

b) As a new node added to be the beginning of the ll

c) As a new node added somewhere in-between the ll.

In turn the ll_add_sorted logic has been updated to use this new add logic.