MCastRecv: 0xffff.ffff to stop
Added logic to server to send 32bit value of 0xffffffff after the full file has
been trasfered, to indicate end of mcast file/data transfer to the clients.

The client logic updated to check for this special stop marker and quit the
mcast recv loop.

Also as part of the same, moved the statistics print logic into a seperate
function which is called periodically as before, as well as when stopping.