QemuBasedTestSetup: Added vm_childs_update
In this case the both the child VMs are run using -net user qemu network stack,
so they both can access the host as well as internet.

Also the qemu's builtin tftp server is enabled for the VMs. The /tmp/share
folder of the host is used as the directory used by the tftp server.
So to share any files to VM, copy them into /tmp/share on the host.

Also the ports 1110 (for VM1) and 1111 (for VM2) of the host are redirected to
ssh ports of the VMs.