UCastRecovery bug fixes
Server side logic correctly handles empty buffer content following the list of
lost packet ranges in the URAckSeqNum packets.

On Client side, As the SeqNum is stored into a signed int rather than unsigned
int, trying to compare if the recieved seqNum is part of the data packets or a
command packet by checking for greater than logic wont work. So have changed
the logic to and the recieved seqNum with a mask which specifies the common
part across all the commands.