UCastRecoverySrvr: Empty LostPackets list handling
A client while sending a URAckSeqNum response with no lost packets to recover,
may send

a) a non-existing application level data part in the packet

OR

b) a application level data part in the packet which is all binary 0's

Previously (a) was being handled, but not (b). Now even (b) is handled.