UnicastRecoverySrvr - better clients list handling
The individual client level unicast recovery logic now returns info as to
whether all the lost packets were successfully recovered for that client or
because of non response from client, it is temporarily giving up, so that one
can try to help recover for other clients.

The main unicast recovery logic now culls the list of clients, if the clients
come back telling no more lost packets for them.

Inturn the main logic will try to help clients recover their lost packets, upto
3 times, even if they had stopped communicating in the middle of error
recovery. However as mentioned earlier, if a client stops responding, other
clients will be given a chance first, before it is retried for this non
responding client.

NOTE: Currently a client not responding back to server for its URReqSeqNum
command, for upto 3 minutes is treated as time out, which indicates that the
client is not recoverable for the current attempt.

The program when exiting gives a proper picture of the current status wrt
content transfer.