UnicastRecovery Server - Max chances per attempt per client
If a client has too many lost packet ranges, such that even after handshaking
with the client for a predefined number of times, still there are lost packets,
then the client will be temporarily kept aside, till other clients are given a
chance once.

To help with this now, the ur_client logic checks how many times it has
handshaked with the client, and then break from its loop if required.

NOTE: This logic is mainly based on how many lost packet ranges are there for a
given client, which is not strictly the same as actual lost packets, but this
keeps things simple for now, while at same time giving a rough estimate of how
many lost packets are there in most cases.

If ur_client is giving up on a client for the current global attempt, then it
prints a message with reason for the same.